# schemas/openapi/query-api.yaml
openapi: 3.0.3
info:
  title: Analytics Query API
  description: Query trends, funnels, retention; dashboards and auth.
  version: 1.0.0

servers:
  - url: http://localhost:8001
    description: Local

paths:
  /health:
    get:
      summary: Liveness
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, properties: { status: { type: string } } }

  /api/trends:
    get:
      summary: Event trend over time
      operationId: getTrends
      security: [{ ApiKeyAuth: [] }]
      parameters:
        - name: project_id
          in: query
          required: true
          schema: { type: string }
        - name: event
          in: query
          required: true
          schema: { type: string }
        - name: date_from
          in: query
          required: true
          schema: { type: string, format: date }
        - name: date_to
          in: query
          required: true
          schema: { type: string, format: date }
        - name: interval
          in: query
          schema: { type: string, enum: [day, week, month], default: day }
      responses:
        "200":
          description: Time series data
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TrendResult" }
        "401":
          description: Unauthorized
        "400":
          description: Bad request

  /api/funnels:
    post:
      summary: Funnel conversion (ordered steps)
      operationId: getFunnels
      security: [{ ApiKeyAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [project_id, steps, date_from, date_to]
              properties:
                project_id: { type: string }
                steps: { type: array, items: { type: string }, minItems: 2 }
                date_from: { type: string, format: date }
                date_to: { type: string, format: date }
      responses:
        "200":
          description: Funnel step counts
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FunnelResult" }
        "401":
          description: Unauthorized
        "400":
          description: Bad request

  /api/query/async:
    post:
      summary: Run heavy query asynchronously
      operationId: runAsyncQuery
      security: [{ ApiKeyAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [project_id, type]
              properties:
                project_id: { type: string }
                type: { type: string, enum: [trend, funnel] }
                params: { type: object }
      responses:
        "202":
          description: Job accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id: { type: string }
                  status: { type: string, example: pending }
        "401":
          description: Unauthorized

  /api/query/async/{job_id}:
    get:
      summary: Get async query result
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: string }
      security: [{ ApiKeyAuth: [] }]
      responses:
        "200":
          description: Result when ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id: { type: string }
                  status: { type: string, enum: [pending, completed, failed] }
                  result: {}
        "202":
          description: Still pending
        "401":
          description: Unauthorized

components:
  schemas:
    TrendResult:
      type: object
      properties:
        series: { type: array, items: { type: object } }
        labels: { type: array, items: { type: string } }
    FunnelResult:
      type: object
      properties:
        steps: { type: array, items: { type: object } }
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization